picsgrabber_src/requirements.txt
Django==1.8.4
django-debug-toolbar==1.4
imgurpython==1.1.7
MySQL-python==1.2.3rc1
requests==2.11.1
sqlparse==0.1.18
facebook-sdk==2.0.0
